unify moe implementation for llama4 and deepseek_v3
