+++ torchtitan/models/deepseek_v3/infra/parallelize.py
@@ -17,7 +17,7 @@ from torch.distributed.tensor.parallel import (
 
 from torchtitan.config import JobConfig, TORCH_DTYPE_MAP
 from torchtitan.distributed import ParallelDims
-from torchtitan.experiments.llama4.infra.expert_parallel import NoParallel
+from torchtitan.distributed.expert_parallel import NoParallel
 from torchtitan.experiments.llama4.infra.parallelize import apply_fsdp, apply_moe_ep_tp
 from torchtitan.models.llama3.infra.parallelize import apply_ac, apply_ddp
 from torchtitan.tools.logging import logger
